+++ addons/im_livechat/static/src/core/public_web/discuss_app_model_patch.js
@@ -6,18 +6,19 @@ import { patch } from "@web/core/utils/patch";
 
 patch(DiscussApp.prototype, {
     setup(env) {
-        super.setup(env);
+        super.setup(...arguments);
         this.defaultLivechatCategory = Record.one("DiscussAppCategory", {
             compute() {
                 return {
                     extraClass: "o-mail-DiscussSidebarCategory-livechat",
-                    icon: "fa fa-commenting-o",
                     hideWhenEmpty: true,
+                    icon: "fa fa-commenting-o",
                     id: `im_livechat.category_default`,
                     name: _t("Livechat"),
                     sequence: 21,
                 };
             },
+            eager: true,
         });
         this.livechats = Record.many("Thread", { inverse: "appAsLivechats" });
     },
