+++ addons/im_livechat/static/src/core/public_web/livechat_channel_model_patch.js
@@ -17,6 +17,7 @@ const livechatChannelPatch = {
                     sequence: 22,
                 };
             },
+            eager: true,
             inverse: "livechatChannel",
         });
         this.threads = Record.many("Thread", { inverse: "livechatChannel" });
