+++ addons/mail/static/src/core/public_web/discuss_app_model.js
@@ -1,35 +1,7 @@
 import { Record } from "@mail/core/common/record";
 import { browser } from "@web/core/browser/browser";
-import { _t } from "@web/core/l10n/translation";
 
 export class DiscussApp extends Record {
-    static new(data) {
-        /** @type {import("models").DiscussApp} */
-        const res = super.new(data);
-        Object.assign(res, {
-            channels: {
-                extraClass: "o-mail-DiscussSidebarCategory-channel",
-                icon: "fa fa-hashtag",
-                id: "channels",
-                name: _t("Channels"),
-                canView: true,
-                sequence: 10,
-                serverStateKey: "is_discuss_sidebar_category_channel_open",
-                addTitle: _t("Add or join a channel"),
-            },
-            chats: {
-                extraClass: "o-mail-DiscussSidebarCategory-chat",
-                icon: "fa fa-users",
-                id: "chats",
-                name: _t("Direct messages"),
-                canView: false,
-                sequence: 30,
-                serverStateKey: "is_discuss_sidebar_category_chat_open",
-                addTitle: _t("Start a conversation"),
-            },
-        });
-        return res;
-    }
     /** @returns {import("models").DiscussApp} */
     static get(data) {
         return super.get(data);
@@ -62,16 +34,7 @@ export class DiscussApp extends Record {
             }
         },
     });
-    allCategories = Record.many("DiscussAppCategory", {
-        inverse: "app",
-        sort: (c1, c2) =>
-            c1.sequence !== c2.sequence
-                ? c1.sequence - c2.sequence
-                : c1.name.localeCompare(c2.name),
-    });
     thread = Record.one("Thread");
-    channels = Record.one("DiscussAppCategory");
-    chats = Record.one("DiscussAppCategory");
     hasRestoredThread = false;
 }
