+++ addons/mail/static/src/core/public_web/thread_model_patch.js
@@ -1,31 +1,10 @@
 import { patch } from "@web/core/utils/patch";
 import { Thread } from "@mail/core/common/thread_model";
-import { Record } from "@mail/core/common/record";
 import { router } from "@web/core/browser/router";
 import { ConfirmationDialog } from "@web/core/confirmation_dialog/confirmation_dialog";
 import { _t } from "@web/core/l10n/translation";
 
 patch(Thread.prototype, {
-    setup() {
-        super.setup(...arguments);
-        this.discussAppCategory = Record.one("DiscussAppCategory", {
-            compute() {
-                return this._computeDiscussAppCategory();
-            },
-        });
-    },
-
-    _computeDiscussAppCategory() {
-        if (this.parent_channel_id) {
-            return;
-        }
-        if (["group", "chat"].includes(this.channel_type)) {
-            return this.store.discuss.chats;
-        }
-        if (this.channel_type === "channel") {
-            return this.store.discuss.channels;
-        }
-    },
     /**
      * Handle the notification of a new message based on the notification setting of the user.
      * Thread on mute:
