+++ addons/mail/static/src/discuss/core/public/discuss_app_model_patch.js
@@ -1,11 +1,12 @@
 import { DiscussApp } from "@mail/core/public_web/discuss_app_model";
+import "@mail/discuss/core/public_web/discuss_app_model_patch";
 
 import { patch } from "@web/core/utils/patch";
 
-patch(DiscussApp, {
-    new() {
-        const res = super.new(...arguments);
-        res.chats.hideWhenEmpty = true;
+patch(DiscussApp.prototype, {
+    computeChats() {
+        const res = super.computeChats(...arguments);
+        res.hideWhenEmpty = true;
         return res;
     },
 });
