+++ addons/mail/static/src/discuss/core/public_web/discuss_app_category_model.js
