+++ addons/mail/static/src/discuss/core/public_web/discuss_patch.js
