+++ addons/mail/static/src/discuss/core/public_web/messaging_menu_patch.js
